Remove string func
